Renamed symbol
