+++ src/main/java/com/simisinc/platform/application/ecommerce/BoxzookaApiClientCommand.java
@@ -27,7 +27,7 @@ import com.fasterxml.jackson.databind.JsonNode;
 import com.github.fge.jackson.JsonLoader;
 import com.simisinc.platform.application.admin.LoadSitePropertyCommand;
 import com.simisinc.platform.application.http.HttpDeleteCommand;
-import com.simisinc.platform.application.http.HttpGetToStringCommand;
+import com.simisinc.platform.application.http.HttpGetCommand;
 import com.simisinc.platform.application.http.HttpPostCommand;
 import com.simisinc.platform.application.http.HttpPutCommand;
 import com.simisinc.platform.domain.model.ecommerce.Order;
@@ -134,7 +134,7 @@ public class BoxzookaApiClientCommand {
     headers.put("Accept", "application/json");
     headers.put("Content-Type", "application/json");
 
-    String remoteContent = HttpGetToStringCommand.execute(url, headers);
+    String remoteContent = HttpGetCommand.execute(url, headers);
 
     // Check for content
     if (StringUtils.isBlank(remoteContent)) {
