+++ src/main/java/com/simisinc/platform/application/ecommerce/TrackingCommand.java
@@ -33,7 +33,7 @@ import org.w3c.dom.NodeList;
 
 import com.simisinc.platform.application.DataException;
 import com.simisinc.platform.application.admin.LoadSitePropertyCommand;
-import com.simisinc.platform.application.http.HttpGetToStringCommand;
+import com.simisinc.platform.application.http.HttpGetCommand;
 
 /**
  * Package Tracking
@@ -76,7 +76,7 @@ public class TrackingCommand {
     try {
       LOG.debug("Sending xml... " + xml);
       String url = USPS_TRACKING_API_URL + URLEncoder.encode(xml, "UTF-8");
-      String responseValue = HttpGetToStringCommand.execute(url);
+      String responseValue = HttpGetCommand.execute(url);
       if (responseValue == null) {
         return null;
       }
