+++ src/main/java/com/simisinc/platform/application/http/HttpPostCommand.java
@@ -38,7 +38,7 @@ import org.apache.commons.validator.routines.UrlValidator;
  */
 public class HttpPostCommand {
 
-  private static Log LOG = LogFactory.getLog(HttpGetToFileCommand.class);
+  private static Log LOG = LogFactory.getLog(HttpDownloadFileCommand.class);
 
   public static final int POST = 1;
   public static final int PATCH = 2;
