+++ src/main/java/com/simisinc/platform/application/http/HttpPutCommand.java
@@ -29,7 +29,7 @@ import org.apache.commons.logging.LogFactory;
  */
 public class HttpPutCommand {
 
-  private static Log LOG = LogFactory.getLog(HttpGetToFileCommand.class);
+  private static Log LOG = LogFactory.getLog(HttpDownloadFileCommand.class);
 
   public static String execute(String url, Map<String, String> parameters) {
     return HttpPostCommand.execute(url, null, parameters, HttpPostCommand.PUT);
